# specs/openapi/platform-v1.yaml
openapi: 3.1.1
info:
  title: Stead Platform API
  version: 0.1.0
  description: Canonical Platform contract. The Phase 1 BFF exposes authorized Stead resources, never provider or infrastructure ontology.
servers:
  - url: /api/v1
...
security:
  - oidc: []
  - localSession: []
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List only individually authorized Organizations
      operationId: listOrganizations
      description: Returns an authorized page, without unfiltered totals or existence metadata. Continuation is the last returned authorized ID, and is emitted only after finding another authorized row. A revoked or cross-scope cursor denies; bounded discovery exhaustion returns a generic failure, never a truncated page.
      parameters: [{$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/After'}]
      responses: {'200': {$ref: '#/components/responses/OrganizationList'}, default: {$ref: '#/components/responses/Problem'}}
    post:
      tags: [Organizations]
      summary: Create an Organization using an explicit instance grant
      operationId: createOrganization
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OrganizationCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/Organization'}, default: {$ref: '#/components/responses/Problem'}}
  /organizations/{organization_id}:
    get:
      tags: [Organizations]
      summary: Get an authorized Organization
      operationId: getOrganization
      parameters: [{$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {$ref: '#/components/responses/Organization'}, default: {$ref: '#/components/responses/Problem'}}
  /organizations/{organization_id}/teams:
    get:
      tags: [Teams]
      summary: List authorized Teams in an Organization
      operationId: listTeams
      description: Hierarchy is returned only after authorization; parentage does not grant access.
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/After'}]
      responses: {'200': {$ref: '#/components/responses/TeamList'}, default: {$ref: '#/components/responses/Problem'}}
    post:
      tags: [Teams]
      summary: Create a Team with explicit permissions and optional structural parent
      operationId: createTeam
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TeamCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/Team'}, default: {$ref: '#/components/responses/Problem'}}
  /teams/{team_id}:
    get:
      tags: [Teams]
      summary: Read an individually authorized Team
      operationId: getTeam
      parameters: [{$ref: '#/components/parameters/TeamId'}]
      responses: {'200': {$ref: '#/components/responses/Team'}, default: {$ref: '#/components/responses/Problem'}}
  /organizations/{organization_id}/projects:
    get:
      tags: [Projects]
      summary: List individually authorized Project views
      operationId: listProjects
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/After'}]
      responses: {'200': {$ref: '#/components/responses/ProjectList'}, default: {$ref: '#/components/responses/Problem'}}
    post:
      tags: [Projects]
      summary: Create a general Project with Work and Docs capabilities
      operationId: createProject
      parameters: [{$ref: '#/components/parameters/OrganizationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/ProjectView'}, default: {$ref: '#/components/responses/Problem'}}
  /projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get an authorization-filtered Project view
      operationId: getProject
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses: {'200': {$ref: '#/components/responses/ProjectView'}, default: {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    oidc: {type: openIdConnect, openIdConnectUrl: '/.well-known/openid-configuration'}
    localSession: {type: apiKey, in: cookie, name: __Host-stead_session, description: Explicit local-development-only bootstrap identity; still subject to the central authorization boundary}
  parameters:
    PageSize: {name: page_size, in: query, description: Maximum authorized items per page; omitted means 20., schema: {type: integer, minimum: 1, maximum: 20}}
    After: {name: after, in: query, description: Last returned authorized resource ID; reauthorized in the requested scope on every use. Not an access token., schema: {$ref: '#/components/schemas/UuidV7'}}
    OrganizationId: {name: organization_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
    TeamId: {name: team_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
    ProjectId: {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8}}
...
  headers:
    ETag: {schema: {type: string}, description: Strong resource version validator}
    SteadSchemaVersion: {schema: {type: string, pattern: '^[1-9][0-9]*\.[0-9]+$'}, description: Exact response representation schema version}
  schemas:
    OrganizationPage:
      type: object
      required: [items]
      properties:
        items: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/Organization'}}
        next_after: {$ref: '#/components/schemas/UuidV7'}
      additionalProperties: false
    TeamPage:
      type: object
      required: [items]
      properties:
        items: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/Team'}}
        next_after: {$ref: '#/components/schemas/UuidV7'}
      additionalProperties: false
    ProjectPage:
      type: object
      required: [items]
      properties:
        items: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/ProjectView'}}
        next_after: {$ref: '#/components/schemas/UuidV7'}
      additionalProperties: false
...
    OrganizationCreate:
      type: object
      required: [key, name]
      properties:
        key: {type: string, pattern: '^[A-Z][A-Z0-9]{1,9}$'}
        name: {type: string, minLength: 1}
      additionalProperties: false
    TeamCreate:
      type: object
      required: [key, name]
      properties:
        key: {type: string, pattern: '^[A-Z][A-Z0-9]{1,9}$'}
        name: {type: string, minLength: 1}
        parent_team_id: {$ref: '#/components/schemas/UuidV7'}
      additionalProperties: false
    ProjectCreate:
      type: object
      required: [key, title, purpose, owning_team_id]
      properties:
        key: {type: string, pattern: '^[A-Z][A-Z0-9]{1,9}$'}
        title: {type: string, minLength: 1}
        purpose: {type: string, minLength: 1}
        owning_team_id: {$ref: '#/components/schemas/UuidV7'}
      additionalProperties: false
    UuidV7: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/UuidV7'}
    CanonicalResourceUri: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/CanonicalResourceUri'}
    CanonicalBrowserUrl: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/CanonicalBrowserUrl'}
    Organization: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/Organization'}
    Team: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/Team'}
...
    ProjectView:
      type: object
      required: [kind, id, uri, browser_url, schema_version, instance_id, scope_kind, scope_id, organization_id, title, purpose, lifecycle_state, owning_team_id, authorized_capabilities, visible_areas, effective_security_label, security_presentation, version]
      properties:
        kind: {const: project}
        id: {$ref: '#/components/schemas/UuidV7'}
        uri: {$ref: '#/components/schemas/CanonicalResourceUri'}
        browser_url: {$ref: '#/components/schemas/CanonicalBrowserUrl'}
        schema_version: {const: '1.0'}
        instance_id: {$ref: '#/components/schemas/UuidV7'}
        scope_kind: {const: organization}
        scope_id: {$ref: '#/components/schemas/UuidV7'}
        organization_id: {$ref: '#/components/schemas/UuidV7'}
        title: {type: string}
        purpose: {type: string}
        lifecycle_state: {enum: [planned, active, paused, completed, canceled]}
        owning_team_id: {$ref: '#/components/schemas/UuidV7'}
        authorized_capabilities: {type: array, items: {enum: [work, docs, scm, code_review, ci, packages, releases, deployments]}, uniqueItems: true}
        visible_areas: {type: array, items: {enum: [overview, work, docs, code, delivery]}, uniqueItems: true}
        effective_security_label: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/SecurityLabelValue'}
        security_presentation: {$ref: '../work-graph-profile/owgp-v0.1.schema.json#/$defs/SecurityPresentation'}
        version: {type: integer, minimum: 1}
      additionalProperties: false
...
    Problem:
      type: object
      required: [type, title, status, correlation_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string, description: Must not disclose denied resource existence or protected metadata}
        instance: {type: string, format: uri-reference}
        correlation_id: {type: string}
      additionalProperties: false
  responses:
...
    OrganizationList: {description: Authorized Organization page only, headers: {Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/OrganizationPage'}}}}
    Team: {description: Authorized Team, headers: {ETag: {$ref: '#/components/headers/ETag'}, Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/Team'}}}}
    ProjectList: {description: Authorized Project page only, headers: {Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectPage'}}}}
    Organization: {description: Organization, headers: {ETag: {$ref: '#/components/headers/ETag'}, Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/Organization'}}}}
    TeamList: {description: Authorized Team page only, headers: {Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/TeamPage'}}}}
    ProjectView: {description: Authorization-filtered Project summary with capability-derived visible areas, headers: {ETag: {$ref: '#/components/headers/ETag'}, Stead-Schema-Version: {$ref: '#/components/headers/SteadSchemaVersion'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectView'}}}}
...
    Problem: {description: RFC 9457 error with non-leaking semantics, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}